refactor
move sync manager to a separate file, making server logic cleaner
